+++ tests/test_central.py
@@ -1,9 +1,15 @@
 """Test the HaHomematic central."""
 from __future__ import annotations
 
+import const
 import helper
-from helper import get_hm_device
+from helper import get_hm_device, load_device_description
 import pytest
+
+TEST_DEVICES: dict[str, str] = {
+    "VCU2128127": "HmIP-BSM.json",
+    "VCU6354483": "HmIP-STHD.json",
+}
 
 
 @pytest.mark.asyncio
@@ -12,9 +18,7 @@ async def test_device_export(
 ) -> None:
     """Test device export."""
     assert central_local_factory
-    central, mock_client = await central_local_factory.get_central(
-        {"VCU6354483": "HmIP-STHD.json"}
-    )
+    central, mock_client = await central_local_factory.get_central(TEST_DEVICES)
     assert central
     assert mock_client
     hm_device = get_hm_device(central_unit=central, address="VCU6354483")
@@ -28,11 +32,107 @@ async def test_all_parameters(
 ) -> None:
     """Test device export."""
     assert central_local_factory
-    central, mock_client = await central_local_factory.get_central(
-        {"VCU6354483": "HmIP-STHD.json"}
-    )
+    central, mock_client = await central_local_factory.get_central(TEST_DEVICES)
     assert central
     assert mock_client
     parameters = central.paramset_descriptions.get_all_readable_parameters()
     assert parameters
-    assert len(parameters) == 26
+    assert len(parameters) == 43
+
+
+@pytest.mark.asyncio
+async def test_add_device(
+    central_local_factory: helper.CentralUnitLocalFactory,
+) -> None:
+    """Test device export."""
+    assert central_local_factory
+    central, mock_client = await central_local_factory.get_central(
+        TEST_DEVICES, ignore_device_on_create=["HmIP-BSM.json"]
+    )
+    assert central
+    assert len(central._devices) == 1
+    assert len(central._entities) == 24
+    assert (
+        len(
+            central.device_descriptions._raw_device_descriptions_persistant_cache.get(
+                const.LOCAL_INTERFACE_ID
+            )
+        )
+        == 9
+    )
+    assert (
+        len(
+            central.paramset_descriptions._raw_paramset_descriptions_persistant_cache.get(
+                const.LOCAL_INTERFACE_ID
+            )
+        )
+        == 8
+    )
+    dev_desc = load_device_description(central=central, filename="HmIP-BSM.json")
+    await central.add_new_devices(const.LOCAL_INTERFACE_ID, dev_desc)
+    assert len(central._devices) == 2
+    assert len(central._entities) == 50
+    assert (
+        len(
+            central.device_descriptions._raw_device_descriptions_persistant_cache.get(
+                const.LOCAL_INTERFACE_ID
+            )
+        )
+        == 20
+    )
+    assert (
+        len(
+            central.paramset_descriptions._raw_paramset_descriptions_persistant_cache.get(
+                const.LOCAL_INTERFACE_ID
+            )
+        )
+        == 18
+    )
+
+
+@pytest.mark.asyncio
+async def test_delete_device(
+    central_local_factory: helper.CentralUnitLocalFactory,
+) -> None:
+    """Test device export."""
+    assert central_local_factory
+    central, mock_client = await central_local_factory.get_central(TEST_DEVICES)
+    assert central
+    assert len(central._devices) == 2
+    assert len(central._entities) == 50
+    assert (
+        len(
+            central.device_descriptions._raw_device_descriptions_persistant_cache.get(
+                const.LOCAL_INTERFACE_ID
+            )
+        )
+        == 20
+    )
+    assert (
+        len(
+            central.paramset_descriptions._raw_paramset_descriptions_persistant_cache.get(
+                const.LOCAL_INTERFACE_ID
+            )
+        )
+        == 18
+    )
+
+    await central.delete_devices(const.LOCAL_INTERFACE_ID, ["VCU2128127"])
+    assert len(central._devices) == 1
+    assert len(central._entities) == 24
+    assert (
+        len(
+            central.device_descriptions._raw_device_descriptions_persistant_cache.get(
+                const.LOCAL_INTERFACE_ID
+            )
+        )
+        == 9
+    )
+    assert (
+        len(
+            central.paramset_descriptions._raw_paramset_descriptions_persistant_cache.get(
+                const.LOCAL_INTERFACE_ID
+            )
+        )
+        == 8
+    )