+++ changelog.md
@@ -5,6 +5,7 @@
 - Remove sleep after connection_checker stops
 - Remove LOWBAT from HM-LC-Sw1-Pl, HM-LC-Sw2-FM
 - Simplify entity de-/registration
+- Refactor add/delete device and add tests
 
 # Version 2023.1.0 (2023-01-01)
 - API Cleanup